apps/ai/pyproject.toml
[tool.poetry]
name = "ai"
version = "1.0.0"
description = "AI service for DEMO-CHAT using FastAPI and Gemini"
authors = ["DEMO-CHAT Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
python-dotenv = "^1.0.0"
google-generativeai = "^0.3.1"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
ruff = "^0.1.9"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
